Move Option as a subcomponent
