+++ src/components/OptionList/components/index.js
@@ -1,0 +1,1 @@
+export { Option } from "./Option";
